@@ -1,10 +1,14 @@
-module github.com/fixme_my_friend/hw10_program_optimization
+module github.com/Tyrqvir/otus_hw/hw10_program_optimization
 
-go 1.16
+go 1.18
+
+require (
+	github.com/goccy/go-json v0.9.11
+	github.com/stretchr/testify v1.7.0
+)
 
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
-	github.com/stretchr/testify v1.7.0
-	gopkg.in/yaml.v2 v2.4.0 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
 	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b // indirect
 )
