@@ -1,21 +1,24 @@
+//go:build !bench
 // +build !bench
 
 package hw10programoptimization
 
 import (
 	"bytes"
+	"io"
+	"reflect"
 	"testing"
 
 	"github.com/stretchr/testify/require"
 )
 
-func TestGetDomainStat(t *testing.T) {
-	data := `{"Id":1,"Name":"Howard Mendoza","Username":"0Oliver","Email":"aliquid_qui_ea@Browsedrive.gov","Phone":"6-866-899-36-79","Password":"InAQJvsq","Address":"Blackbird Place 25"}
+var data = `{"Id":1,"Name":"Howard Mendoza","Username":"0Oliver","Email":"aliquid_qui_ea@Browsedrive.gov","Phone":"6-866-899-36-79","Password":"InAQJvsq","Address":"Blackbird Place 25"}
 {"Id":2,"Name":"Jesse Vasquez","Username":"qRichardson","Email":"mLynch@broWsecat.com","Phone":"9-373-949-64-00","Password":"SiZLeNSGn","Address":"Fulton Hill 80"}
 {"Id":3,"Name":"Clarence Olson","Username":"RachelAdams","Email":"RoseSmith@Browsecat.com","Phone":"988-48-97","Password":"71kuz3gA5w","Address":"Monterey Park 39"}
 {"Id":4,"Name":"Gregory Reid","Username":"tButler","Email":"5Moore@Teklist.net","Phone":"520-04-16","Password":"r639qLNu","Address":"Sunfield Park 20"}
 {"Id":5,"Name":"Janice Rose","Username":"KeithHart","Email":"nulla@Linktype.com","Phone":"146-91-01","Password":"acSBF5","Address":"Russell Trail 61"}`
 
+func TestGetDomainStat(t *testing.T) {
 	t.Run("find 'com'", func(t *testing.T) {
 		result, err := GetDomainStat(bytes.NewBufferString(data), "com")
 		require.NoError(t, err)
@@ -36,4 +39,141 @@ func TestGetDomainStat(t *testing.T) {
 		require.NoError(t, err)
 		require.Equal(t, DomainStat{}, result)
 	})
+}
+
+func Test_countDomains(t *testing.T) {
+	type args struct {
+		u users
+	}
+	tests := []struct {
+		name    string
+		args    args
+		want    DomainStat
+		wantErr bool
+	}{
+		{
+			name: "Test count domains",
+			args: args{
+				u: []User{
+					{
+						ID:       0,
+						Name:     "",
+						Username: "",
+						Email:    "test1@biz.com",
+						Phone:    "",
+						Password: "",
+						Address:  "",
+					},
+					{
+						ID:       0,
+						Name:     "",
+						Username: "",
+						Email:    "test1@bit.com",
+						Phone:    "",
+						Password: "",
+						Address:  "",
+					},
+				},
+			},
+			want: DomainStat{
+				"biz.com": 1,
+				"bit.com": 1,
+			},
+			wantErr: false,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			got, err := countDomains(tt.args.u)
+			if (err != nil) != tt.wantErr {
+				t.Errorf("countDomains() error = %v, wantErr %v", err, tt.wantErr)
+				return
+			}
+			if !reflect.DeepEqual(got, tt.want) {
+				t.Errorf("countDomains() got = %v, want %v", got, tt.want)
+			}
+		})
+	}
+}
+
+func Test_getUsers(t *testing.T) {
+	type args struct {
+		r      io.Reader
+		domain string
+	}
+	tests := []struct {
+		name       string
+		args       args
+		wantResult users
+		wantErr    bool
+	}{
+		{
+			name: "Test com domain",
+			args: args{
+				r:      bytes.NewBufferString(data),
+				domain: "com",
+			},
+			wantResult: []User{
+				{
+					ID:       2,
+					Name:     "Jesse Vasquez",
+					Username: "qRichardson",
+					Email:    "mLynch@broWsecat.com",
+					Phone:    "9-373-949-64-00",
+					Password: "SiZLeNSGn",
+					Address:  "Fulton Hill 80",
+				},
+				{
+					ID:       3,
+					Name:     "Clarence Olson",
+					Username: "RachelAdams",
+					Email:    "RoseSmith@Browsecat.com",
+					Phone:    "988-48-97",
+					Password: "71kuz3gA5w",
+					Address:  "Monterey Park 39",
+				},
+				{
+					ID:       5,
+					Name:     "Janice Rose",
+					Username: "KeithHart",
+					Email:    "nulla@Linktype.com",
+					Phone:    "146-91-01",
+					Password: "acSBF5",
+					Address:  "Russell Trail 61",
+				},
+			},
+			wantErr: false,
+		},
+		{
+			name: "Test gov domain",
+			args: args{
+				r:      bytes.NewBufferString(data),
+				domain: "gov",
+			},
+			wantResult: []User{
+				{
+					ID:       1,
+					Name:     "Howard Mendoza",
+					Username: "0Oliver",
+					Email:    "aliquid_qui_ea@Browsedrive.gov",
+					Phone:    "6-866-899-36-79",
+					Password: "InAQJvsq",
+					Address:  "Blackbird Place 25",
+				},
+			},
+			wantErr: false,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			gotResult, err := getUsers(tt.args.r, tt.args.domain)
+			if (err != nil) != tt.wantErr {
+				t.Errorf("getUsers() error = %v, wantErr %v", err, tt.wantErr)
+				return
+			}
+			if !reflect.DeepEqual(gotResult, tt.wantResult) {
+				t.Errorf("getUsers() gotResult = %v, want %v", gotResult, tt.wantResult)
+			}
+		})
+	}
 }
